requirements.txt
Pillow==9.4.0
Pillow==10.1.0
streamlit==1.27.2
